individualize each autocmd
