+++ lua/core/autocmd/autosaved.lua
@@ -1,0 +1,4 @@
+vim.api.nvim_create_autocmd(
+	{ "FocusLost", "ModeChanged", "TextChanged", "BufEnter" },
+	{ desc = "autosave", pattern = "*", command = "silent! update" }
+)
