+++ lua/core/autocmd/init.lua
@@ -1,0 +1,3 @@
+require("core.autocmd.autosaved")
+require("core.autocmd.lspconfig")
+require("core.autocmd.git")
