+++ lua/core/autocmd/lspconfig.lua
@@ -1,8 +1,3 @@
-vim.api.nvim_create_autocmd(
-	{ "FocusLost", "ModeChanged", "TextChanged", "BufEnter" },
-	{ desc = "autosave", pattern = "*", command = "silent! update" }
-)
-
 vim.api.nvim_create_autocmd("LspAttach", {
 	group = vim.api.nvim_create_augroup("kickstart-lsp-attach", { clear = true }),
 	callback = function(event)
@@ -41,25 +36,6 @@ vim.api.nvim_create_autocmd("LspAttach", {
 			map("<leader>th", function()
 				vim.lsp.inlay_hint.enable(not vim.lsp.inlay_hint.is_enabled({ bufnr = event.buf }))
 			end, "[T]oggle Inlay [H]ints")
-		end
-	end,
-})
-
-vim.api.nvim_create_autocmd("BufReadPost", {
-	group = vim.api.nvim_create_augroup("GitProjectDetection", { clear = true }),
-	callback = function(event)
-		local function is_git_repo(path)
-			path = path or vim.fn.expand("%:p:h")
-			while path ~= "/" do
-				if vim.fn.isdirectory(path .. "/.git") == 1 then
-					return true
-				end
-				path = vim.fn.fnamemodify(path, ":h")
-			end
-			return false
-		end
-		if is_git_repo() then
-			require("core.keymaps_autocmd").load_plugins("lazygit")
 		end
 	end,
 })
